+++ src/main/java/de/eldoria/companies/data/wrapper/order/OrderContent.java
@@ -16,7 +16,7 @@ public class OrderContent {
     private ItemStack stack;
     private int amount;
     private List<ContentPart> parts = new ArrayList<>();
-    private float price;
+    private double price;
 
     public OrderContent(ItemStack stack, int amount, float price) {
         this.stack = stack;
@@ -40,7 +40,7 @@ public class OrderContent {
         return parts;
     }
 
-    public float price() {
+    public double price() {
         return price;
     }
 
@@ -78,10 +78,10 @@ public class OrderContent {
         return amount - delivered();
     }
 
-    public Map<UUID, Float> payments() {
-        Map<UUID, Float> payments = new HashMap<>();
+    public Map<UUID, Double> payments() {
+        Map<UUID, Double> payments = new HashMap<>();
         for (var part : parts) {
-            payments.put(part.worker, (part.amount / (float) amount) * price);
+            payments.put(part.worker, (part.amount / (double) amount) * price);
         }
         return payments;
     }
